+++ src/main/java/com/dragovorn/dragonbot/api/bot/file/FileManager.java
@@ -68,7 +68,7 @@ public class FileManager {
             } catch (IOException exception) {
                 exception.printStackTrace();
             }
-        }); // TESTME
+        });
     }
 
     @Nullable
