+++ src/main/java/com/dragovorn/dragonbot/bot/BotConfiguration.java
@@ -37,7 +37,6 @@ public class BotConfiguration extends Configuration {
         this.defaults.put("channel", "");
         this.defaults.put("console", true);
         this.defaults.put("auto connect", false);
-        this.defaults.put("snapshot versions", false);
         this.defaults.put("ask for update", true);
         this.defaults.put("check for updates", true);
         this.defaults.put("twitch-api key", ""); // This might get removed.
@@ -61,10 +60,6 @@ public class BotConfiguration extends Configuration {
 
     public boolean getAutoConnect() {
         return getBoolean("auto connect");
-    }
-
-    public boolean getPreReleases() {
-        return getBoolean("snapshot versions");
     }
 
     public boolean getAskForUpdates() {
@@ -93,10 +88,6 @@ public class BotConfiguration extends Configuration {
 
     public void setAutoConnect(boolean autoConnect) {
         set("auto connect", autoConnect);
-    }
-
-    public void setPreReleases(boolean preReleases) {
-        set("snapshot versions", preReleases);
     }
 
     public void setAskForUpdates(boolean askForUpdates) {
